smlrc: fix gcc warning
/home/stas/src/SmallerC/v0100/smlrc.c: In function 'IncludeFile':
/home/stas/src/SmallerC/v0100/smlrc.c:1160:11: warning: 'strcpy' offset [-2147483647, -2] is out of the bounds [0, 768] of object 'FileNames' with type 'char[8][96]' [-Warray-bounds=]
 1160 |           strcpy(FileNames[FileCnt] + plen + 1, TokenValueString);
      |           ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/home/stas/src/SmallerC/v0100/smlrc.c:586:6: note: 'FileNames' declared here
  586 | char FileNames[MAX_INCLUDES][MAX_FILE_NAME_LEN + 1];
      |      ^~~~~~~~~

strlen() returns size_t so gcc assumes int can overflow.

+++ v0100/smlrc.c
@@ -1153,7 +1153,7 @@ void IncludeFile(int quot)
       }
       for (i = 0; i < pl; )
       {
-        int plen = strlen(paths + i);
+        size_t plen = strlen(paths + i);
         if (plen + 1 + nlen < MAX_FILE_NAME_LEN)
         {
           strcpy(FileNames[FileCnt], paths + i);
